refactor(cloud-function): rename handlers
