+++ cloud-function/src/handlers/handle-stripe-plans.ts
@@ -15,7 +15,10 @@ interface Result {
   plans: PlanResult;
 }
 
-export async function stripePlans(req: express.Request, res: express.Response) {
+export async function handleStripePlans(
+  req: express.Request,
+  res: express.Response
+) {
   const lookupData =
     ORIGIN_MAIN === "developer.mozilla.org" ? prodLookup : stageLookup;
