+++ cloud-function/src/handlers/proxy-api.ts
@@ -2,8 +2,8 @@ import { createProxyMiddleware, fixRequestBody } from "http-proxy-middleware";
 
 import { Source, sourceUri } from "../env.js";
 
-export const proxyRumba = createProxyMiddleware({
-  target: sourceUri(Source.rumba),
+export const proxyApi = createProxyMiddleware({
+  target: sourceUri(Source.api),
   changeOrigin: true,
   autoRewrite: true,
   proxyTimeout: 20000,
