+++ cloud-function/src/handlers/proxy-content.ts
