+++ cloud-function/src/handlers/proxy-telemetry.ts
